requirements.txt
torch>=2.2.0
torchaudio>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
tqdm>=4.66.0
